refactor: split node.rs helpers into focused submodules (archive, cache, common, env, feedback, git_utils, logging)
