+++ crates/core-node-internal/src/services/node/cache.rs
@@ -1,0 +1,13 @@
+//! Persistent on-disk caches shared by the batch node-add pipeline.
+//!
+//! `git` caches fully-cloned checkouts keyed by `(repo_url, ref)`;
+//! `bundle` caches extracted HTTP archives keyed by `(url, sha256)`.
+//! Both use `key` (slug + short hash) to produce deterministic directory
+//! names under [`config::consts::PeppyDirs`].
+
+pub(super) mod bundle;
+pub(super) mod git;
+mod key;
+
+pub(super) use bundle::ensure_bundle;
+pub(super) use git::ensure_checkout;
