+++ crates/core-node-internal/src/services/node/init/templates.rs
